+++ internal/utils.go
@@ -1,7 +1,7 @@
 package vm
 
-func extendSign(value int, bits int) int {
-	mask := 1 << (bits - 1)
+func extendSign(value int32, bits int) int32 {
+	mask := int32(1 << (bits - 1))
 	if mask&value != 0 {
 		return value | (^(mask - 1))
 	}
